// packages/server/src/entities/enums/PostAwardStage.ts
export enum PostAwardStage {
  ORDER_FOLLOWUP = 'ORDER_FOLLOWUP',
  LOA_PROCESSING = 'LOA_PROCESSING',
  ORDER_PROCESSING = 'ORDER_PROCESSING',
  INSPECTION = 'INSPECTION',
  DISPATCH_DELIVERY = 'DISPATCH_DELIVERY',
  WARRANTY = 'WARRANTY',
  BILL_PAYMENT = 'BILL_PAYMENT',
  SD_RETURN = 'SD_RETURN',
  CLOSED = 'CLOSED',
}